Edge Computing/backend/requirements.txt
flask
Flask-Cors
pymongo
paho-mqtt
requests